requirements.txt
PyMuPDF
pytesseract
pdf2image
Pillow
google-generativeai
gspread
oauth2client
openpyxl
streamlit